[dxva] Add execution fence for the separate decoder device / shared buffers path
On recent AMD there are often flashes / wrong picture when skipping in 4k video due to the lack of
synchronization between the main d3d device and the device of the decoder (when separate).

The GPU fence is set right after the picture decode or copy is queued. When the picture
is needed for rendering, the CPU will wait for the GPU to reach the fence before submitting
work using the picture.