[videodb] Preserve path hash when turning a movie into a version
This is a small optimization: the path hash is deleted when a movie is turned into a version, and
during the next libray scan, the folder of the movie is scanned (unnecessary) and the same hash
is recreated.
The file of the version hasn't moved or changed, the hash is still valid, might as well not delete
it in the first place and save the directory scan.